% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{align_to_range}
\alias{align_to_range}
\title{align_to_range}
\usage{
align_to_range(p, s, cut)
}
\arguments{
\item{p}{A base sequence in character format}

\item{s}{A base sequence in character format}

\item{cut}{The distance between the starting sites of two fragments}
}
\value{
A list include two IRanges instances (deletion and insertion)
}
\description{
Title
}
\examples{
align_to_range(p="AAGG---AAATTTCGGAATAAGGAATTT",s="AAGGCCCCAAATTT-CGGAATAAGGAATTT",cut=0)


}
