% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{BuildTree}
\alias{BuildTree}
\title{BuildTree}
\usage{
BuildTree(tag)
}
\arguments{
\item{tag}{List generate from TagProcess, for more see \code{\link[LinTInd]{TagProcess}}}
}
\value{
list with two elements, a data.tree data structure and a dataframe of array information for each cell barcode
}
\description{
Generate an array generant tree of a data.tree data structure and save it
}
\details{
Title
}
\examples{
data("example_data",package = "LinTInd")
treeinfo<-BuildTree(tag)

}
