% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{change_form_stat}
\alias{change_form_stat}
\title{change_form_stat}
\usage{
change_form_stat(indel)
}
\arguments{
\item{indel}{List include two IRanges instances, contain start and end site of deletions and inserstions}
}
\value{
A scar form string
}
\description{
Title
}
\examples{
data("example_data",package = "LinTInd")
change_form_stat(cellsinfo$indel[[1]])


}
