% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{FindIndel}
\alias{FindIndel}
\title{FindIndel}
\usage{
FindIndel(
  data,
  scarfull,
  scar,
  align_score = NULL,
  type = NULL,
  indel.coverage = NULL,
  cln
)
}
\arguments{
\item{data}{data frame, include cell barcode, UMI and reads.}

\item{scarfull}{DNAString of reference sequence}

\item{scar}{The cutsite data frame}

\item{align_score}{The minimum alignment score that matched sequence should get, default in this parameter is the score that the reads which all of the target set were cutted got}

\item{type}{Group name for this data ("None" in default)}

\item{indel.coverage}{Choose indels selected scope: "Accurate" (default) means only the indles happenned in target site will be idented; "All" means each indel will be detected even they locate on the anchors}

\item{cln}{The number of threads}
}
\value{
list include IRanges instances (deletion and insertion), a data frame of reads' informations, reference sequenc, dataframe of cut sites
}
\description{
This function can ident indels for each reads in input data, and create IRanges instances for deletion and insertion
}
\details{
Title
}
\examples{
data("example_data",package = "LinTInd")
scarinfo<-FindIndel(data=data,scarfull=ref,scar=cutsite,indel.coverage="All",type="test",cln=8)


}
