% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{IndelForm}
\alias{IndelForm}
\title{IndelForm}
\usage{
IndelForm(scarinfo, cln)
}
\arguments{
\item{scarinfo}{List generate from FindIndel, for more see \code{\link[LinTInd]{FindIndel}}}

\item{cln}{The number of threads}
}
\value{
A new list of scarinfo, the scarform of each reads will add in the data frame of reads' informations
}
\description{
Generate scar form strings from scarinfo list for each reads
}
\details{
Title
}
\examples{
data("example_data",package = "LinTInd")
IndelForm(scarinfo,cln=4)

}
