% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{IndelIdents}
\alias{IndelIdents}
\title{IndelIdents}
\usage{
IndelIdents(scarinfo, method.use = NULL, cln)
}
\arguments{
\item{scarinfo}{List generate from IndelForm, for more see \code{\link[LinTInd]{IndelForm}}}

\item{method.use}{Select how to determine a scar form string for each cell:
"reads.num" (default):find the scar with the most reads in the cell；
"umi.num":find the scar with the most UMIs in the cell；
"consensus":find the consistent sequences in each cell, and then generate scar form strings from the new reads}

\item{cln}{The number of threads}
}
\value{
The list generate from FindIndel, but in 'Scar' element a new column contain scar form strings
}
\description{
Function to define a scarform for each cell(single cell) or each reads(bulk seq, generate 'cell barcode' for each reads)
}
\details{
Title
}
\examples{
data("example_data",package = "LinTInd")
IndelIdents(scarinfo,method.use="umi.num",cln=4)


}
