% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{IndelPlot}
\alias{IndelPlot}
\title{IndelPlot}
\usage{
IndelPlot(cellsinfo)
}
\arguments{
\item{cellsinfo}{List generate from IndelIdents, for more see \code{\link[LinTInd]{IndelIdents}}}
}
\value{
2 line charts
}
\description{
Return 2 line charts, show the probability of insertion and deletion at each site
}
\details{
Title
}
\examples{
data("example_data",package = "LinTInd")
IndelPlot(cellsinfo = cellsinfo)

}
