% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{ReadFasta}
\alias{ReadFasta}
\title{ReadFasta}
\usage{
ReadFasta(filename)
}
\arguments{
\item{filename}{The input fasta file name}
}
\value{
A DNAString object
}
\description{
Function to read fasta file to DNAString object
}
\details{
Title
}
\examples{
fafile=paste0(system.file("extdata",package = 'LinTInd'),"/V3.fasta")
ReadFasta(fafile)

}
