% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{TagDist}
\alias{TagDist}
\title{TagDist}
\usage{
TagDist(tag, method = NULL)
}
\arguments{
\item{tag}{List generate from TagProcess, for more see \code{\link[LinTInd]{TagProcess}}}

\item{method}{Denote which method to use:
"Jaccard"(default): calculate the weighted jaccard similarity of indels between each pair of groups;
"P": right-tailed test, compare the Indels intersection level with the hypothetical  result generated from random editing, and the former is expected to be significantly higher than the latter
"spearman": Spearman correlation of indels between each pair of groups}
}
\value{
2 figures are saved to show the distribution of INDEL and the relationship between groups respectively, the matrix of the relationship between groups is returned
}
\description{
If the cell barcode and the anntation of each cell are provided, this function can calculate the relationship between each cell type in three way
}
\details{
Title
}
\examples{
tag_dist=TagDist(tag,method = "spearman")
tag_dist=TagDist(tag)
tag_dist=TagDist(tag,method = "Jaccard")
tag_dist=TagDist(tag,method = "P")
tag_dist=TagDist(tag,method = "spearman")

}
