% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{TagProcess}
\alias{TagProcess}
\title{TagProcess}
\usage{
TagProcess(data, Cells = NULL, prefix = NULL)
}
\arguments{
\item{data}{List generate from IndelIdents, for more see \code{\link[LinTInd]{IndelIdents}}}

\item{Cells}{(optional) Dataframe of cells' annotation, with two columns: "Cell.BC" and "Cell.type"}

\item{prefix}{(optional) Indels' prefix}
}
\value{
List with two dataframes: Indels for each cell barcode and cells' annotation
}
\description{
Split each indel from scar form strings and map indel information to cell barcodes
}
\details{
Title
}
\examples{
data("example_data",package = "LinTInd")
TagProcess(cellsinfo$info,Cells=celltype)

}
