% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{PlotTree}
\alias{PlotTree}
\title{PlotTree}
\usage{
PlotTree(treeinfo, data.extract = NULL, annotation = NULL, prefix = NULL)
}
\arguments{
\item{treeinfo}{List generate from BuildTree, for more see \code{\link[LinTInd]{BuildTree}}}

\item{data.extract}{(optional) If "F" (default), will not return the indel's information, if it's "T", the opposite}

\item{annotation}{(optional) If "T" (default), the annotation of each cell barcodes have to be provided before, and a heatmap of cells' distribution for each array will be return}

\item{prefix}{(optional) Indels' prefix}
}
\value{

}
\description{
Function to visualise the array generant tree
}
\details{
Title
}
\examples{
data("example_data",package = "LinTInd")
Shows the distribution of each cell type
plotinfo<-PlotTree(treeinfo = treeinfo,data.extract = "T",annotation = "T")
Only show the tree and indel patterns
plotinfo<-PlotTree(treeinfo = treeinfo,data.extract = "T",annotation = "F")

}
